package restructuring, day/night effects added, taxation added, tweaks to dialogue priority, shifts added for work venues

+++ src/game/building/Inventory.java
@@ -74,7 +74,7 @@ public class Inventory {
   
   /**  Financial balance-
     */
-  public void incCredits(int inc) {
+  public void incCredits(float inc) {
     if (inc > 0) {
       credits += inc ;
     }
